kronos/requirements.txt
psycopg2-binary
sentence-transformers
numpy
dotenv
